+++ tests/testthat/test_feature_plot.R
@@ -1,11 +1,12 @@
 test <- readRDS("../test_data/test_object.rds")
+test_data <- test$dex$test
 
 test_that("plot_features returns correct object for one gene", {
-  d <- test$data |>
+  d <- test_data$data |>
     dplyr::filter(id == "G1") |>
     dplyr::left_join(test$features, by = "id") |>
-    dplyr::left_join(test$meta, by = "sample")
-  plt <- plot_features(d)
+    dplyr::left_join(test_data$metadata, by = "sample")
+  plt <- plot_features(d, what = "value", x_var = "group")
   expect_true(ggplot2::is.ggplot(plt))
 
   data_used <- ggplot2::ggplot_build(plt)$data[[1]]
@@ -16,14 +17,14 @@ test_that("plot_features returns correct object for one gene", {
 
 
 test_that("plot_features returns correct object for two genes", {
-  d <- test$data |>
+  d <- test_data$data |>
     dplyr::left_join(test$features, by = "id") |>
-    dplyr::left_join(test$meta, by = "sample")
-  plt <- plot_features(d)
+    dplyr::left_join(test_data$metadata, by = "sample")
+  plt <- plot_features(d, what = "value", x_var = "sample")
   expect_true(ggplot2::is.ggplot(plt))
 
   dm <- d |>
-    dplyr::left_join(test$meta, by = "sample")
+    dplyr::left_join(test_data$metadata, by = "sample")
 
   data_used <- ggplot2::ggplot_build(plt)$data[[1]]
   expect_equal(data_used$x |> as.numeric(), dm$sample |> as.factor() |> as.numeric())
@@ -32,7 +33,7 @@ test_that("plot_features returns correct object for two genes", {
 
 
 test_that("plot_features returns NULL when data are empty", {
-  d <-  test$data |>
+  d <-  test_data$data |>
     dplyr::filter(id == "none")
   plt <- plot_features(d)
   expect_null(plt)